Refactor dev- og playwright-oppsett - men beholder msw for test
