+++ apps/ungdomsytelse-veileder/.env.example
@@ -7,6 +7,5 @@ UNG_DELTAKELSE_OPPLYSER_FRONTEND_PATH=/api/ung-deltakelse-opplyser
 UNG_DELTAKELSE_OPPLYSER_API_SCOPE=api://dev-gcp.k9saksbehandling.ung-deltakelse-opplyser/.default
 UNG_DELTAKELSE_OPPLYSER_API_URL=http://ung-deltakelse-opplyser.k9saksbehandling
 
-DEV_USE_MSW=false
 SIF_PUBLIC_USE_FARO=true
 SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL=http://localhost:12347/collect
