+++ apps/ungdomsytelse-veileder/.storybook/preview-head.html
@@ -6,7 +6,6 @@
         'PUBLIC_PATH': '/',
         'GITHUB_REF_NAME': 'dev',
         'DEV_IS_STORYBOOK': true,
-        'DEV_USE_MSW': true,
         'SIF_PUBLIC_USE_FARO': true,
         'SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL': '#',
