+++ apps/ungdomsytelse-veileder/env.schema.ts
@@ -7,10 +7,6 @@ const localSchema = z.object({
         .optional()
         .transform((val) => (typeof val === 'boolean' ? val : val === 'true'))
         .optional(),
-    DEV_USE_MSW: z
-        .union([z.boolean(), z.string()])
-        .optional()
-        .transform((val) => (typeof val === 'boolean' ? val : val === 'true')),
     SIF_PUBLIC_USE_FARO: z
         .union([z.boolean(), z.string()])
         .optional()
