+++ apps/ungdomsytelse-veileder/mock/data/registrertDeltakerMock.ts
