+++ apps/ungdomsytelse-veileder/mock/msw/browser.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers/handlers';
+import { handlers } from './handlers';
 
 export const worker = setupWorker(...handlers);
