+++ apps/ungdomsytelse-veileder/mock/msw/enableMocking.ts
@@ -1,0 +1,7 @@
+export async function enableMocking() {
+    if (import.meta.env.MODE !== 'msw') {
+        return;
+    }
+    const { worker } = await import('./browser');
+    return worker.start({ onUnhandledRequest: 'bypass' });
+}
