+++ apps/ungdomsytelse-veileder/nais/prod-gcp.json
@@ -30,7 +30,6 @@
         "UNG_DELTAKELSE_OPPLYSER_API_SCOPE": "api://prod-gcp.k9saksbehandling.ung-deltakelse-opplyser/.default",
         "UNG_DELTAKELSE_OPPLYSER_API_URL": "http://ung-deltakelse-opplyser.k9saksbehandling",
 
-        "DEV_USE_MSW": false,
         "SIF_PUBLIC_USE_FARO": true,
         "SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL": "https://telemetry.nav.no/collect"
     }
