+++ apps/ungdomsytelse-veileder/src/forms/endre-periode-form/EndreStartdato.stories.tsx
@@ -6,7 +6,7 @@ import EndrePeriodeForm from './EndrePeriodeForm';
 import { withModalWrapper } from '../../../storybook/decorators/withModalWrapper';
 import { withQueryClientProvider } from '../../../storybook/decorators/withQueryClientProvider';
 import { EndrePeriodeVariant } from '../../types/EndrePeriodeVariant';
-import { registrertDeltakerMock } from '../../../mock/msw/mocks/data/registrertDeltakerMock';
+import { registrertDeltakerMock } from '../../../mock/data/registrertDeltakerMock';
 import { deltakelseSchema } from '../../types/Deltakelse';
 import { registrertDeltakerSchema } from '../../types/Deltaker';
