+++ apps/ungdomsytelse-veileder/src/pages/start-page/StartPage.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { BrowserRouter } from 'react-router-dom';
 import { http, HttpResponse } from 'msw';
-import { nyDeltakerMock } from '../../../mock/msw/mocks/data/nyDeltakerMock';
+import { nyDeltakerMock } from '../../../mock/data/nyDeltakerMock';
 import { withDarkBg } from '../../../storybook/decorators/withDarkBg';
 import { withIntl } from '../../../storybook/decorators/withIntl';
 import { withPageWidth } from '../../../storybook/decorators/withPageWidth';
