+++ apps/ungdomsytelse-veileder/tsconfig.json
@@ -8,7 +8,8 @@
     "exclude": ["node_modules", "src/build", "**.js", "dist"],
     "include": [
         "./src/**/*",
-        "./mock/msw/**/*",
+        "./mock/**/*",
+        "./playwright/**/*",
         "./vite-env.d.ts",
         "./storybook/**/*",
         "./vite.config.ts",
