+++ apps/ungdomsytelse-veileder/vite.dev.config.ts
@@ -15,7 +15,6 @@ const appSettings: AppEnv = {
     UNG_DELTAKELSE_OPPLYSER_API_URL: 'http://localhost:8089',
     UNG_DELTAKELSE_OPPLYSER_API_SCOPE: 'dev-gcp:dusseldorf:ung-deltakelse-opplyser',
 
-    DEV_USE_MSW: true,
     SIF_PUBLIC_USE_FARO: true,
     SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL: 'http://localhost:12347/collect',
 };
@@ -23,6 +22,7 @@ const appSettings: AppEnv = {
 const veileder = { name: 'Pål', NAVident: 'Z999999' };
 
 export default defineConfig({
+    mode: 'msw',
     plugins: [
         tailwindcss(),
         react({
